- implementacao stats

+++ src/main/java/application/rest/v1/VerificacoesWrapper.java
@@ -1,26 +1,26 @@
 package application.rest.v1;
 
 public class VerificacoesWrapper {
-	private int count_mutant_dna;
-	private int count_human_dna;
-	private int ratio;
+	private double count_mutant_dna;
+	private double count_human_dna;
+	private double ratio;
 	
-	public int getCount_mutant_dna() {
+	public double getCount_mutant_dna() {
 		return count_mutant_dna;
 	}
-	public void setCount_mutant_dna(int count_mutant_dna) {
+	public void setCount_mutant_dna(double count_mutant_dna) {
 		this.count_mutant_dna = count_mutant_dna;
 	}
-	public int getCount_human_dna() {
+	public double getCount_human_dna() {
 		return count_human_dna;
 	}
-	public void setCount_human_dna(int count_human_dna) {
+	public void setCount_human_dna(double count_human_dna) {
 		this.count_human_dna = count_human_dna;
 	}
-	public int getRatio() {
+	public double getRatio() {
 		return ratio;
 	}
-	public void setRatio(int ratio) {
+	public void setRatio(double ratio) {
 		this.ratio = ratio;
 	}
 }
